+++ app/src/main/java/com/abdeveloper/library/MultiSelectModelExtra.java
@@ -1,4 +1,4 @@
-package io.github.getsixtyfour.openpyn.utilities;
+package com.abdeveloper.library;
 
 import androidx.annotation.NonNull;
